+++ structures/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "structures.apps.StructuresConfig"
 
-__version__ = "1.5.4"
+__version__ = "1.6.0b1"
 __title__ = "Structures"